[BUGFIX] Fixed sorting of albums in FE
The sorting column was already present but was not set in
TCA as sorting column. This changes the default sys_file_collection
a little as the sorting is now enabled.

Sorting is first on sorting column ASC and then on crdate DESC

Updated TCA so the Override available since TYPO 6.2.2