[FEATURE] Nested view now shows all selected albums
When albums are selected for the nested album view all albums are
shown now even when they are not connected to root level. For instance
if rootlevel album isn't allowed.

All albums without acceseble parent are seen as root level so shown on
first album list.

A new variable in fluid is introduced here parentAlbum so you know
when to link to the parent or to the album list.

Fixes: #18